### 2021-08-06 #3583 @whedon
[PRE REVIEW]: obscura: A modular C++ tool and library for the direct detection of (sub-GeV) dark matter via nuclear and electron recoils
**Submitting author:** @temken (<a href="http://orcid.org/0000-0002-4251-2229">Timon Emken</a>)
**Repository:** <a href="https://github.com/temken/obscura" target ="_blank">https://github.com/temken/obscura</a>
**Version:** v0.1.1
**Reviewers:** @rafaelab @villaa
