### 2021-08-08 22cdc0c
Update DOI to v0.1.1
