+++ README.md
@@ -5,7 +5,8 @@
 
 # obscura - Direct detection of dark matter with nucleus and electron recoil experiments
 
-[![DOI](https://zenodo.org/badge/250298432.svg)](https://zenodo.org/badge/latestdoi/250298432)
+[![status](https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9/status.svg)](https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5163944.svg)](https://doi.org/10.5281/zenodo.5163944)
 
 A modular C++ tool and library for dark matter direct detection computations for both nuclear and electron recoil experiments.
 
@@ -17,7 +18,7 @@ You can find more detailed documentation of *obscura* [here](https://obscura.rea
 
 If you decide to use this code, or if you want to add a reference to it, please cite the latest archived version,
 
-> Emken, T., 2021, obscura - A C++ library for dark matter detection computations [Code, v0.1.0] [[DOI:10.5281/zenodo.4557188]](https://zenodo.org/record/4557188).
+> Emken, T., 2021, obscura - A C++ library for dark matter detection computations [Code, v0.1.1] [[DOI:10.5281/zenodo.5163944]](https://zenodo.org/record/5163944).
 
 <details><summary>Bibtex entry</summary>
 <p>
@@ -28,10 +29,10 @@ If you decide to use this code, or if you want to add a reference to it, please 
   title = {{obscura - A C++ library for dark matter detection computations [Code, v0.1.0]}},
   year         = {2021},
   publisher    = {Zenodo},
-  version      = {v0.1.0},
-  doi          = {DOI:10.5281/zenodo.4557188},
-  url          = {https://doi.org/10.5281/zenodo.4557188},
-  howpublished={The code can be found under \url{https://github.com/temken/obscura}. Version 0.1.0 is archived as \href{https://doi.org/10.5281/zenodo.4557188}{DOI:10.5281/zenodo.4557188}}
+  version      = {v0.1.1},
+  doi          = {DOI:10.5281/zenodo.5163944},
+  url          = {https://doi.org/10.5281/zenodo.5163944},
+  howpublished={The code can be found under \url{https://github.com/temken/obscura}. Version 0.1.1 is archived as \href{https://doi.org/10.5281/zenodo.5163944}{DOI:10.5281/zenodo.5163944}}
 }
 ```
 
@@ -40,6 +41,7 @@ If you decide to use this code, or if you want to add a reference to it, please 
 
 ## VERSION HISTORY
 
+- 05.08.2021: Release of version 0.1.1
 - 23.02.2021: Release of version 0.1.0
 
 ## AUTHORS & CONTACT
