### 2021-08-12 5d3a2e9
Add function to check if total cross sections depend on relative speed of DM-target
