+++ src/DM_Particle_Standard.cpp
@@ -320,6 +320,15 @@ double DM_Particle_SI::d2Sigma_dq2_dEe_Crystal(double q, double Ee, double vDM, 
 }
 
 //Total cross sections
+
+bool DM_Particle_SI::Is_Sigma_Total_V_Dependent() const
+{
+	if(!low_mass || FF_DM == "General")
+		return true;
+	else
+		return false;
+}
+
 double DM_Particle_SI::Sigma_Total_Nucleus(const Isotope& isotope, double vDM, double param)
 {
 	double sigmatot = 0.0;
@@ -528,6 +537,11 @@ double DM_Particle_SD::dSigma_dq2_Electron(double q, double vDM, double param) c
 }
 
 //Total cross sections with nuclear isotopes, elements, and electrons
+bool DM_Particle_SD::Is_Sigma_Total_V_Dependent() const
+{
+	return false;
+}
+
 double DM_Particle_SD::Sigma_Total_Nucleus(const Isotope& isotope, double vDM, double param)
 {
 	return (isotope.spin != 0) ? 4.0 * pow(libphysica::Reduced_Mass(mass, isotope.mass), 2.0) / M_PI * (isotope.spin + 1.0) / isotope.spin * pow(fp * isotope.sp + fn * isotope.sn, 2.0) : 0.0;
