### 2021-08-23 #3583 @danielskatz
👋 @lucydot - do you think you could edit this submission for JOSS?
