### 2021-09-13 #3725 @whedon
[REVIEW]: obscura: A modular C++ tool and library for the direct detection of (sub-GeV) dark matter via nuclear and electron recoils
**Submitting author:** @temken (<a href="http://orcid.org/0000-0002-4251-2229">Timon Emken</a>)
**Repository:** <a href="https://github.com/temken/obscura" target ="_blank">https://github.com/temken/obscura</a>
**Version:** v1.0.0

@rafaelab @villaa, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
