### 2021-10-04 #3725 @rafaelab
I finished testing the code. It was easy to install and to execute. Everything worked as expected and the additional tests I did show that the package is robust. The only thing I really missed was a folder with the cpp files corresponding to the examples shown in the docs.

The last point is that there are no guidelines for people wishing to contribute to the software, as required by JOSS.

I'm happy to recommend `obscura' for publication once these minor issues and the ones about the paper are addressed by @temken.
