+++ paper/paper.md
@@ -23,7 +23,7 @@ This *dark matter* (DM) must be fundamentally different from the known matter we
 Its only established property is that it interacts gravitationally and indeed dominates the gravitational potential of galaxies and galaxy clusters.
 One of the leading hypothesis is that DM is made up of one or more new particles and that galaxies such as our Milky Way are embedded in gigantic haloes of these as of yet undetected particles.
 Our planet would at any moment be penetrated by a stream of these particles without much of an effect.
-If these dark particles interact with nuclei and/or electrons via some new force besides gravity, they would on occasion collide with a terrestrial atom.
+If these dark particles interact with nuclei and/or electrons via some new force besides gravity, they would on occasion collide with a terrestrial particle.
 *Direct detection experiments* search for these kind of interactions and aim to observe DM events within a detector caused by an interaction with target nuclei [@Goodman:1984dc;@Drukier:1986tm;@Wasserman:1986hh] or electrons [@Kopp:2009et;@Essig:2011nj].
 These experiments are typically placed deep underground to shield them from possible backgrounds e.g. due to cosmic rays.
 
@@ -41,7 +41,7 @@ For more details on `obscura` and its implementation in C++, we refer to the [do
 
 [^1]: The latest version of the documentation can be found under [https://obscura.readthedocs.io](https://obscura.readthedocs.io).
 
-![Excluded regions (95% confidence level) of the DM parameter space given by the $(m_\mathrm{DM},\sigma_i)$ plane, where $m_\mathrm{DM}$ is the assumed DM mass and $\sigma_i$ is the interaction cross section with target $i$. \label{fig:constraints}](obscura_DD_Constraints.png){ width=70% } 
+![Excluded regions (95% confidence level) of the DM parameter space given by the $(m_\mathrm{DM},\sigma_i)$ plane, where $m_\mathrm{DM}$ is the assumed DM mass and $\sigma_i$ is the interaction cross section with target $i$. For comparison, the dashed lines denote the official results published by the experimental collaborations. Note that some of the collaborations present their results with 90% CL. Some of the `obscura` results are conservative due to a simplified analysis.\label{fig:constraints}](obscura_DD_Constraints.png){ width=70% } 
 
 
 
@@ -57,7 +57,7 @@ As an example, let us look at the energy spectrum of DM induced ionization event
 The DM mass $m_\mathrm{DM}$ and the differential DM-electron scattering cross section $\frac{\mathrm{d}\sigma_e}{\mathrm{d}q^2}$ are defined by the assumed particle physics of the hypothetical DM particle the experiment is probing.
 The velocity distribution $f_\chi(v)$ and the local DM energy density $\rho_\chi$ are important inputs from astrophysics and cosmology.
 Lastly, the ionization form factor $f_\mathrm{ion}^{n\ell}(q,E_e)$ encapsulates the atomic physics of the electronic bound states and describes the probability of an electron with quantum numbers $(n\ell)$ to get ionized by an incoming DM particle.
-As we can see, the evaluation this expression for the electron recoil spectrum is highly modular combining inputes from various fields of research.
+As we can see, the evaluation of this expression for the electron recoil spectrum is highly modular combining inputs from various fields of research.
 This modularity should be reflected in the structure of corresponding research software.
 
 It is our ambition for the `obscura` code that the basic functionality does not rely on specific choices and that no particular assumption is hard-coded.
