### 2021-10-06 7c1a964
JOSS Review: Add examples from the documentation as cpp files
