### 2021-10-06 849cc24
JOSS Review: Add guidelines for potential contributors
