### 2021-10-06 9653cf1
JOSS Review: Add clarifications on which DM particles can be treated.
