### 2021-10-06 #3725 @temken
Hi @rafaelab,
Thank you for reviewing the paper and code, the positive feedback, and the constructive suggestions. I will respond to each issue raised and summarize the changes I made. All changes have just been merged into the [main branch](https://github.com/temken/obscura).

## 1. Regarding the paper

>     * l. 18: "terrestrial atom". I'd use something else broader than just "atom". Something like "terrestrial particle".
>
>     * l. 53: "the evaluation this expression" => missing "of"
>
>     * l. l. 54: inputes => inputs

I implemented these changes in [paper/paper.md](https://github.com/temken/obscura/blob/master/paper/paper.md)

[[Link to the corresponding commit]](https://github.com/temken/obscura/commit/7933e93a320b9a786710c3783f313dd33c3eef00)

> As for the content, the text does not specify that DM particles that can be treated with the code. Sure, the code is flexible enough to enable the user to input their favourite model. But there are certainly some limitations. Would it work, for example, with heavier particles, for example (at least the ones that can penetrate the atmosphere and reach the detector)? To be fair, this is written in the statement of need, which I personally find a bit too late for the reader.

I added a few clarifying statements to the third paragraph of the Summary in [paper/paper.md](https://github.com/temken/obscura/blob/master/paper/paper.md).

[[Link to the corresponding commit]](https://github.com/temken/obscura/commit/9653cf1271451c51ad9df9c1fdf8fee351ab77fa)

> I'm also missing how the results obtained with `obscura` compare with other other published results. I suppose other people have already produced figures like figure 1. If this is the case, how do the curves compares with yours?

This is a very good idea. I have added a dashed line to figure 1 to indicate the official results as reported by the experimental collaborations. This was possible for most experiments with the exception of SD nuclear interactions. I also extended the figure description.

[[Link to the corresponding commit 1]](https://github.com/temken/obscura/commit/7933e93a320b9a786710c3783f313dd33c3eef00)
[[Link to the corresponding commit 2]](https://github.com/temken/obscura/commit/e7a15aec7ffd46867a63206082cd70ec70b2e1ce)

## 2. Regarding the software

> I finished testing the code. It was easy to install and to execute. Everything worked as expected and the additional tests I did show that the package is robust. The only thing I really missed was a folder with the cpp files corresponding to the examples shown in the docs.

All examples found in the [example page](https://obscura.readthedocs.io/en/latest/07_Examples.html) of the documentation are now also found as cpp files in [obscura/examples](https://github.com/temken/obscura/tree/master/examples). I have also tested them again, and they compile and run successfully.

[[Link to the corresponding commit]](https://github.com/temken/obscura/commit/7c1a9643d6de585a42a745c90e222c790984623f)

> The last point is that there are no guidelines for people wishing to contribute to the software, as required by JOSS.

Thank you for this nice suggestion. I added the file [docs/CONTRIBUTING.md](https://github.com/temken/obscura/tree/master/docs/CONTRIBUTING.md) to clarify to everyone how they can contribute.

[[Link to the corresponding commit]](https://github.com/temken/obscura/commit/849cc24aed9dc92a458e4ab9c379c77152f2495c)

> I'm happy to recommend `obscura' for publication once these minor issues and the ones about the paper are addressed by @temken.

Thanks again for the valuable feedback! I hope the code/paper is now suitable for publication in JOSS.
