### 2021-10-27 ab9e6b1
JOSS Review: CONTRIBUTING.md linked to documentation
