### 2021-10-28 97e607b
JOSS Review: Remove redundant information from reference
