+++ paper/paper.bib
@@ -95,7 +95,7 @@
   version      = {v0.1.0},
   doi          = {DOI:10.5281/zenodo.4559874},
   url          = {https://doi.org/10.5281/zenodo.4559874},
-  howpublished={Astrophysics Source Code Library record \href{https://ascl.net/2102.018}{[ascl:2102.018]}. The code can be found under \url{https://github.com/temken/damascus-sun}. Version 0.1.0 is archived as \href{https://doi.org/10.5281/zenodo.4559874}{DOI:10.5281/zenodo.4559874}}
+  howpublished={Astrophysics Source Code Library record \href{https://ascl.net/2102.018}{[ascl:2102.018]}. The code can be found under \url{https://github.com/temken/damascus-sun}.}
 }
 
 @article{Essig:2011nj,
