### 2021-10-28 51f2d90
Fix libphysica version for 0.2.0 release
