+++ CMakeLists.txt
@@ -24,7 +24,7 @@ include(FetchContent)
 set(LIBPHYSICA_DIR   ${EXTERNAL_DIR}/libphysica)
 FetchContent_Declare(libphysica
   GIT_REPOSITORY  https://github.com/temken/libphysica.git
-  GIT_TAG         origin/master
+  GIT_TAG         v0.1.1
   SOURCE_DIR     "${LIBPHYSICA_DIR}"
   CMAKE_ARGS     "-DCMAKE_BUILD_TYPE=Release -DCODE_COVERAGE=OFF" )
 FetchContent_GetProperties(libphysica)
