### 2021-10-28 5ada9ea
Update zenodo DOI of v0.2.0
