+++ docs/source/index.rst
@@ -15,8 +15,8 @@
 *obscura* - Direct detection of dark matter with nucleus and electron recoil experiments
 ========================================================================================
 
-.. image:: https://zenodo.org/badge/250298432.svg
-   :target: https://zenodo.org/badge/latestdoi/250298432
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5607961.svg
+   :target: https://doi.org/10.5281/zenodo.5607961
    :alt: DOI
 
 A modular C++ tool and library for dark matter direct detection computations for both nuclear and electron recoil experiments.
