+++ docs/source/09_Citations.rst
@@ -8,7 +8,7 @@ How to cite
 
 If you decide to use this code, or if you want to add a reference to it, please cite the latest archived version,
 
-    Emken, T., 2021, obscura - A C++ library for dark matter detection computations [Code, v0.1.1] [DOI:10.5281/zenodo.5163944]
+    Emken, T., 2021, obscura - A C++ library for dark matter detection computations [Code, v0.2.0] [DOI:10.5281/zenodo.5607961]
 
 .. raw:: html
 
@@ -19,12 +19,12 @@ If you decide to use this code, or if you want to add a reference to it, please 
 
     @software{obscura,
     author = {Emken, Timon},
-    title = {{obscura - A C++ library for dark matter detection computations [Code, v0.1.0]}},
+    title = {{obscura - A C++ library for dark matter detection computations [Code, v0.2.0]}},
     year         = {2021},
     publisher    = {Zenodo},
-    version      = {v0.1.1},
-    doi          = {DOI:10.5281/zenodo.5163944},
-    url          = {https://doi.org/10.5281/zenodo.5163944},
+    version      = {v0.2.0},
+    doi          = {DOI:10.5281/zenodo.5607961},
+    url          = {https://doi.org/10.5281/zenodo.5607961},
     howpublished={The code can be found under \url{https://github.com/temken/obscura}.}
     }
 
