### 2021-10-28 67baa38
Add JOSS badge to documentation
