+++ docs/source/index.rst
@@ -18,6 +18,9 @@
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5607961.svg
    :target: https://doi.org/10.5281/zenodo.5607961
    :alt: DOI
+.. image:: https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9/status.svg
+   :target: https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9
+   :alt: JOSS paper
 
 A modular C++ tool and library for dark matter direct detection computations for both nuclear and electron recoil experiments.
 
