### 2021-10-29 4e433f4
Improve the generated version header by distinguishing project and top level.
