### 2021-11-02 #3725 @lucydot
Hi @temken I've just realised the version is 0.2.0 - I'd usually expect production/stable code to be at a version number 1.0.0+ (with 0.X.Y for beta releases). Was this an oversight per chance?
