### 2021-11-02 #3725 @temken
Hi @lucydot,
I didn't really think much about it, it was just that all previous version were 0.1.x, so I thought 0.2.0 would be the next step. I can make another release with version 1.0.0 if necessary. Please let me know.
