### 2021-11-08 #3725 @lucydot
Hi @temken - I'd consider releasing this as 1.0.0. It's not a deal-breaker, I'll leave the final decision up to you - it may be best to follow what is typically done in your particular domain. However in my experience 0.X.Y versions are those used for initial development  - see, for example, the guidance for semantic versioning: https://semver.org/.
