+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.12)
 
-project(obscura VERSION "0.2.0")
+project(obscura VERSION "1.0.0")
 
 # Use C++-11
 set(CMAKE_CXX_STANDARD 11)
