+++ docs/source/index.rst
@@ -15,8 +15,8 @@
 *obscura* - Direct detection of dark matter with nucleus and electron recoil experiments
 ========================================================================================
 
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5607961.svg
-   :target: https://doi.org/10.5281/zenodo.5607961
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5665890.svg
+   :target: https://doi.org/10.5281/zenodo.5665890
    :alt: DOI
 .. image:: https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9/status.svg
    :target: https://joss.theoj.org/papers/fd8076268036956d3bf08193c4fc2db9
