### 2021-11-10 #3725 @temken
Hi @lucydot,
Thank you for the link, which was very instructive. I'll definitely be more mindful on my versioning in the future. I just released version 1.0.0 (tag (v1.0.0)[https://github.com/temken/obscura/releases/tag/v1.0.0]) replacing the previous v0.2.0, which I deleted.  Version 1.0.0 is archived under [DOI:10.5281/zenodo.5665890](https://zenodo.org/record/5665890).

I also made a very minor change to the acknowledgments of the paper (specification of a grant).
