### 2021-11-12 #3725 @kyleniemeyer
Hi @temken, it looks like this submission is essentially at the finish line, so I'm going to step in and help wrap it up.

I found a few minor issues in the paper to fix:
- Throughout the article, can you please add commas after "e.g.", and commas before if used in the sentence (for example, at the end of the first paragraph, it should be "... backgrounds, e.g., due to cosmic rays.")
- Before equation (1), I recommend changing it to `... as derived by @Essig:2015cda`, which will make it "as derived by Essig et al. (2015)".
- In the statement of need, I would remove the "see e.g." you use in two places, and just have the example citations at the end of the sentences.
