### 2021-11-13 #3725 @temken
Hi @kyleniemeyer,

Thanks a lot for looking over the paper. I implemented the changes you suggested and also ran a spell-check.

[Link to commit](https://github.com/temken/obscura/pull/35/commits/52c37bc5adaaaa6c7513b0aeaa939cc9e0232b17)

I hope it is now ready for publication. If not let me know :)
Best regards,
Timon
