### 2021-11-13 #3725 @temken
Since the last changes were only minor fixes to the paper, could we use the DOI provided before (v1.0.0)? [10.5281/zenodo.5665890](https://zenodo.org/record/5665890)

Or is it necessary to make a new release and generate a new DOI (v1.0.1)?
