### 2021-11-21 #3725 @lucydot
@temken it looks like the pre-print from Griffin et al has just been published in PRD (10.1103/physrevd.104.095015) - please could you update the paper reference to reflect this? Thank-you
