+++ paper/paper.bib
@@ -192,7 +192,11 @@
     eprint = "2105.05253",
     archivePrefix = "arXiv",
     primaryClass = "hep-ph",
-    month = "5",
+    doi = "10.1103/PhysRevD.104.095015",
+    journal = "Phys. Rev. D",
+    volume = "104",
+    number = "9",
+    pages = "095015",
     year = "2021"
 }
 
