### 2021-11-30 #3725 @kthyng
Hi @temken! I'll help finish publishing your paper. First, can you update the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper?
